% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R, R/yacc.R
\name{dbg}
\alias{dbg}
\title{Print debug message.}
\usage{
dbg(msg)

dbg(msg)
}
\arguments{
\item{msg}{message to display.}

\item{msg}{message to display.}
}
\description{
Print debug message.

Print debug message.
}

