% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{get_regex}
\alias{get_regex}
\title{-----------------------------------------------------------------------------
get_regex(func)}
\usage{
get_regex(func)
}
\description{
Returns the regular expression assigned to a function.
-----------------------------------------------------------------------------
}

