% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{Grammar}
\alias{Grammar}
\title{-----------------------------------------------------------------------------
                          === GRAMMAR CLASS ===}
\format{An \code{\link{R6Class}} generator object}
\usage{
Grammar
}
\description{
The following class represents the contents of the specified grammar along
with various computed properties such as first sets, follow sets, LR items, etc.
This data is used for critical parts of the table generation process later.
-----------------------------------------------------------------------------
}
\keyword{data}

