% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\name{id}
\alias{id}
\title{Unique identificator of environment object.}
\usage{
id(x)
}
\arguments{
\item{x}{environment}
}
\value{
The address of the object
}
\description{
This function is retrieving an phisical address for an 
environment object using R internal inspect function
}

