% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{data}
\name{is_identifier}
\alias{is_identifier}
\title{-----------------------------------------------------------------------------
                         === Grammar Representation ===}
\format{An object of class \code{character} of length 1.}
\usage{
is_identifier
}
\description{
The following functions, classes, and variables are used to represent and
manipulate the rules that make up a grammar.
-----------------------------------------------------------------------------
regex matching identifiers
}
\keyword{datasets}

