% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{class}
\name{Lexer}
\alias{Lexer}
\title{Lexing Engine}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
Lexer
}
\description{
{
The following Lexer class implements the lexer runtime. There are only
a few public methods and attributes:

\itemize{
 \item input() - Store a new string in the lexer
 \item token() - Get the next token
 \item clone() - Clone the lexer

 \item lineno  - Current line number
 \item lexpos  - Current position in the input string
}
}
}
\keyword{datasets}
