% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{class}
\name{LexerReflect}
\alias{LexerReflect}
\title{LexerReflect
This class represents information needed to build a lexer as extracted from a
user's input file.}
\format{An \code{\link{R6Class}} generator object}
\usage{
LexerReflect
}
\description{
LexerReflect
This class represents information needed to build a lexer as extracted from a
user's input file.
}
\keyword{data}

