% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{data}
\name{LexToken}
\alias{LexToken}
\title{Token class.  This class is used to represent the tokens produced.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LexToken
}
\description{
Token class.  This class is used to represent the tokens produced.
}
\keyword{datasets}

