% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{LRGeneratedTable}
\alias{LRGeneratedTable}
\title{LRGeneratedTable}
\format{An \code{\link{R6Class}} generator object}
\usage{
LRGeneratedTable
}
\description{
This class implements the LR table generation algorithm.  There are no
public methods
-----------------------------------------------------------------------------
}
\keyword{data}

