% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{LRItem}
\alias{LRItem}
\title{-----------------------------------------------------------------------------
class LRItem}
\format{An \code{\link{R6Class}} generator object}
\usage{
LRItem
}
\description{
This class represents a specific stage of parsing a production rule.  For
example:
}
\details{
expr : expr . PLUS term

In the above, the "." represents the current location of the parse.  Here
basic attributes:

name       - Name of the production.  For example 'expr'
      prod       - A list of symbols on the right side ['expr','.', 'PLUS','term']
      number     - Production number.

lr_next      Next LR item. Example, if we are ' expr -> expr . PLUS term'
                   then lr_next refers to 'expr -> expr PLUS . term'
      lr_index   - LR item index (location of the ".") in the prod list.
      lookaheads - LALR lookahead symbols for this item
      len        - Length of the production (number of symbols on right hand side)
      lr_after    - List of all productions that immediately follow
      lr_before   - Grammar symbol immediately before
-----------------------------------------------------------------------------
}
\keyword{data}

