% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{LRParser}
\alias{LRParser}
\title{== LRParser ==
The LR Parsing engine.}
\format{An \code{\link{R6Class}} generator object}
\usage{
LRParser
}
\description{
== LRParser ==
The LR Parsing engine.
}
\keyword{data}

