% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{LRTable}
\alias{LRTable}
\title{-----------------------------------------------------------------------------
                           == Class LRTable ==}
\format{An \code{\link{R6Class}} generator object}
\usage{
LRTable
}
\description{
This basic class represents a basic table of LR parsing information.
Methods for generating the tables are not defined here.  They are defined
in the derived class LRGeneratedTable.
-----------------------------------------------------------------------------
}
\keyword{data}

