% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{NullLogger}
\alias{NullLogger}
\title{Null logger is used when no output is generated. Does nothing.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NullLogger
}
\description{
Null logger is used when no output is generated. Does nothing.
}
\keyword{datasets}

