% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\name{parse_grammar}
\alias{parse_grammar}
\title{Parse user's grammar}
\usage{
parse_grammar(name, doc)
}
\arguments{
\item{name}{Name of the grammar method}

\item{doc}{Rule description}
}
\value{
grammar
}
\description{
This takes a raw grammar rule string and parses it into production data
}

