% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{ParserReflect}
\alias{ParserReflect}
\title{ParserReflect()}
\format{An \code{\link{R6Class}} generator object}
\usage{
ParserReflect
}
\description{
This class represents information extracted for building a parser including
start symbol, error function, tokens, precedence list, action functions,
etc.
-----------------------------------------------------------------------------
}
\keyword{datasets}

