% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{Production}
\alias{Production}
\title{-----------------------------------------------------------------------------
class Production:}
\format{An \code{\link{R6Class}} generator object}
\usage{
Production
}
\description{
This class stores the raw information about a single production or grammar rule.
A grammar rule refers to a specification such as this:
}
\details{
expr : expr PLUS term

Here are the basic attributes defined on all productions

name     - Name of the production.  For example 'expr'
      prod     - A list of symbols on the right side ['expr','PLUS','term']
      prec     - Production precedence level
      number   - Production number.
      func     - Function that executes on reduce

The following attributes are defined or optional.

len       - Length of the production (number of symbols on right hand side)
      usyms     - Set of unique symbols found in the production
-----------------------------------------------------------------------------
}
\keyword{data}

