% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\name{randomString}
\alias{randomString}
\title{Generate random string.}
\usage{
randomString(length=12)
}
\arguments{
\item{length}{expected number of characters}
}
\value{
Random string
}
\description{
This function generates a random string 
of a given length only from letters
}

