% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{data}
\name{reg_is_identifier}
\alias{reg_is_identifier}
\title{This regular expression is used to match valid token names}
\format{An object of class \code{character} of length 1.}
\usage{
reg_is_identifier
}
\description{
This regular expression is used to match valid token names
}
\keyword{datasets}

