% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\name{rightmost_terminal}
\alias{rightmost_terminal}
\title{-----------------------------------------------------------------------------
rightmost_terminal()}
\usage{
rightmost_terminal(symbols, terminals)
}
\description{
Return the rightmost terminal from a list of symbols.  Used in add_production()
-----------------------------------------------------------------------------
}

