% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rly-package.R
\docType{package}
\name{rly-package}
\alias{rly}
\alias{rly-package}
\title{rly: Tools to Create Formal Language Parsers}
\description{
R implementation of the common parsing tools 'lex' and 'yacc'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/systemincloud/rly}
  \item Report bugs at \url{https://github.com/systemincloud/rly/issues}
}

}
\author{
\strong{Maintainer}: Marek Jagielski \email{marek.jagielski@gmail.com} [copyright holder]

Authors:
\itemize{
  \item David M. Beazley [copyright holder]
}

Other contributors:
\itemize{
  \item Yasutaka Tanaka [contributor]
  \item Henrico Witvliet [contributor]
}

}
\keyword{internal}
