% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{RlyLogger}
\alias{RlyLogger}
\title{This object is a stand-in for a logging object created by the
logging module.   RLY will use this by default to create things
such as the parser.out file.  If a user wants more detailed
information, they can create their own logging object and pass
it into RLY.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RlyLogger
}
\description{
This object is a stand-in for a logging object created by the
logging module.   RLY will use this by default to create things
such as the parser.out file.  If a user wants more detailed
information, they can create their own logging object and pass
it into RLY.
}
\keyword{datasets}

