% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{statetoken}
\alias{statetoken}
\title{def _statetoken(s,names)}
\usage{
statetoken(s, names)
}
\description{
Given a declaration name s of the form "t_" and a dictionary whose keys are
state names, this function returns a tuple (states,tokenname) where states
is a tuple of state names and tokenname is the name of the token.  For example,
calling this with s = "t_foo_bar_SPAM" might return (('foo','bar'),'SPAM')
}

