% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\name{yacc}
\alias{yacc}
\title{Build a parser}
\usage{
yacc(module = NA, args = list(), method = "LALR", debug = FALSE,
  start = NA, check_recursion = TRUE, debugfile = "parser.out",
  outputdir = NA, debuglog = NA, errorlog = NA)
}
\description{
Build a parser
}

