% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{YaccSymbol}
\alias{YaccSymbol}
\title{Non-terminal grammar symbol}
\format{An \code{\link{R6Class}} generator object}
\usage{
YaccSymbol$new()
}
\description{
This class is used to hold non-terminal grammar symbols during parsing.
It normally has the following attributes set:
\itemize{
 \item type      - Grammar symbol type
 \item value     - Symbol value
 \item lineno    - Starting line number
 \item endlineno - Ending line number (optional, set automatically)
 \item lexpos    - Starting lex position
 \item endlexpos - Ending lex position (optional, set automatically)
}
}
\keyword{datasets}

