% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{lex}
\alias{lex}
\title{Build all of the regular expression rules from definitions in the supplied module}
\usage{
lex(module = NA, args = list(), debug = FALSE)
}
\description{
Build all of the regular expression rules from definitions in the supplied module
}

