% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{class}
\name{Lexer}
\alias{Lexer}
\title{Lexing Engine}
\format{An \code{\link{R6Class}} generator object}
\usage{
Lexer
}
\description{
The following Lexer class implements the lexer runtime. There are only
a few public methods and attributes:
  input()          -  Store a new string in the lexer
  token()          -  Get the next token
  clone()          -  Clone the lexer
}
\details{
lineno           -  Current line number
  lexpos           -  Current position in the input string
}
\keyword{data}

