% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{class}
\name{Lexer}
\alias{Lexer}
\title{=== Lexing Engine ===
The following Lexer class implements the lexer runtime. There are only
a few public methods and attributes:
  input()          -  Store a new string in the lexer
  token()          -  Get the next token
  clone()          -  Clone the lexer}
\format{An \code{\link{R6Class}} generator object}
\usage{
Lexer
}
\description{
lineno           -  Current line number
  lexpos           -  Current position in the input string
}
\keyword{-}
\keyword{------------------------------------------------------------}
\keyword{Changes}
\keyword{Lexer}
\keyword{Push}
\keyword{Read}
\keyword{Restores}
\keyword{Return}
\keyword{Returns}
\keyword{Skip}
\keyword{Write}
\keyword{a}
\keyword{ahead}
\keyword{and}
\keyword{begin()}
\keyword{characters}
\keyword{current}
\keyword{current_state()}
\keyword{data}
\keyword{file}
\keyword{from}
\keyword{information}
\keyword{input()}
\keyword{into}
\keyword{lexer}
\keyword{lexing}
\keyword{n}
\keyword{new}
\keyword{next}
\keyword{old}
\keyword{on}
\keyword{opttoken()}
\keyword{pop_state()}
\keyword{previous}
\keyword{push_state()}
\keyword{readtab()}
\keyword{saves}
\keyword{skip()}
\keyword{stack}
\keyword{state}
\keyword{string}
\keyword{tab}
\keyword{table}
\keyword{the}
\keyword{to}
\keyword{token}
\keyword{writetab()}

