% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{class}
\name{LexToken}
\alias{LexToken}
\title{Lex Token}
\format{An \code{\link{R6Class}} generator object}
\usage{
LexToken
}
\description{
Token class.  This class is used to represent the tokens produced
}
\keyword{datasets}

