% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{NullLogger}
\alias{NullLogger}
\title{Null logger is used when no output should be generated.}
\format{A \code{\link{R6Class}} object}
\usage{
NullLogger
}
\description{
Does nothing.
}
\examples{
debuglog <- NullLogger$new()
debuglog$info('This will not print')
}
\keyword{datasets}

