% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\name{parse_grammar}
\alias{parse_grammar}
\title{-----------------------------------------------------------------------------
                           === INTROSPECTION ===}
\usage{
parse_grammar(name, doc)
}
\description{
The following functions and classes are used to implement the PLY
introspection features followed by the yacc() function itself.
-----------------------------------------------------------------------------
-----------------------------------------------------------------------------
parse_grammar()
}
\details{
This takes a raw grammar rule string and parses it into production data
-----------------------------------------------------------------------------
}

