% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{RlyLogger}
\alias{RlyLogger}
\title{Print log message to file or console.}
\format{A \code{\link{R6Class}} object}
\usage{
RlyLogger
}
\description{
This object is a stand-in for a logging object created by the
logging module. RLY will use this by default to create things
such as the parser.out file. If a user wants more detailed
information, they can create their own logging object and pass
it into RLY.
'
}
\examples{
debuglog <- rly::RlyLogger$new(".", "file.out")
debuglog$info('This is info message')
}
\keyword{datasets}
