% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacc.R
\docType{class}
\name{YaccSymbol}
\alias{YaccSymbol}
\title{This class is used to hold non-terminal grammar symbols during parsing.
It normally has the following attributes set:
       .type       = Grammar symbol type
       .value      = Symbol value
       .lineno     = Starting line number
       .endlineno  = Ending line number (optional, set automatically)
       .lexpos     = Starting lex position
       .endlexpos  = Ending lex position (optional, set automatically)}
\format{An \code{\link{R6Class}} generator object}
\usage{
YaccSymbol
}
\description{
This class is used to hold non-terminal grammar symbols during parsing.
It normally has the following attributes set:
       .type       = Grammar symbol type
       .value      = Symbol value
       .lineno     = Starting line number
       .endlineno  = Ending line number (optional, set automatically)
       .lexpos     = Starting lex position
       .endlexpos  = Ending lex position (optional, set automatically)
}
\keyword{data}

